logthrdestdrv: split worker and worker_instance structs
With this change `worker' would contain methods (e.g. static data associated
with the worker thread) and `worker_instance' contains dynamic data
(such as connected and the queue).

This patch will scatter the code with references to the static
"worker_instance", which would eventually be removed by followups when the
LogThreadedDestWorker and LogThreadedDestDriver split is complete.  The
worker_instance would remain long term at least until there are drivers that
don't get updated to the multi-threaded interface.  Once that is complete,
worker_instance can be removed.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>