logthrdestdrv: move methods to LogThreadedDestWorker
A previous patch has started to move the data members to the
LogThreadedDestWorker, this patch moves methods so their
self argument refers to LogThreadedDestWorker. There's still
a circular dependency between the two objects (Driver and Worker), but
that can be eliminated later.

This one actually compiles and runs but _assumes_ that there's
only a single thread running and there's no functionality to
spawn more than one.

Also, the prototype of the derived functions (e.g. insert, flush, etc)
has not been changed.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>